buttons, string-entry
Wrote a new plugin for using the html canvas with buttons. Wrote a plugin for doing string entry. Changed previous functions to use buttons instead of keystrokes. Wrote new functions for forward and backward digit span recall. Wrote a new function for the stroop task. Started writing a new plugin for processing speed: string comparison. This one will take some finagling in order to get the timer to work across trials. What I think needs to happen is instead of doing the trials in a loop, I need to write a plugin that does all the trials sequentially, using the same timer.