Finished tasks.
The last thing I did was finish up the distrator task (which took a heckuva lot longer than expected, I might add). Then I made some changes to the tasks so that the engine can send them lists of stimuli instead of just one at a time. This results in a cleaner interface. Awaiting Dave's approval on task presentation, but it's pretty darn close to Taylor's implementation, just with some of the transitions smoothed out to prevent frustration (especially RE: the distractors).